Fix: Resolve 688 compilation errors (Issue #169) - Phase 1 Progress

+++ ai-coach-api/src/services/model_training_service.rs
@@ -266,7 +266,7 @@ impl ModelTrainingService {
             user_id,
             prediction_type: "ModelMetrics".to_string(),
             data: prediction_data,
-            confidence: Some(metrics.r_squared),
+            confidence: Some(metrics.r_squared as f64),
             model_version: Some(metrics.model_version.clone()),
         };
